It works at this point. I am making this commit so we could do multiple branches

+++ nanolib2.0/tensor/tensor_util.hpp
@@ -3,8 +3,9 @@
 
 #include "absl/container/inlined_vector.h"
 #include "tensor_types.hpp"
+#include "eigen/Eigen/Core"
 
-using floatX = int;
+using floatX = float;
 
 // Raw pointer -> Flat
 template <typename T>
